Template setup for new homepage design

+++ wp-content/themes/source-tech/header.php
@@ -45,6 +45,11 @@
             'phone_conversion_number': '800-932-0657'
         });
     </script>
+    
+    <!-- Font Awesome CDN -->
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css" integrity="sha384-50oBUHEmvpQ+1lW4y57PTFmhCaXp0ML5d60M1M7uH2+nqUivzIebhndOJK28anvf" crossorigin="anonymous">
+
+
 
 
 </head>
@@ -135,9 +140,14 @@
 	</header><!-- #masthead -->
 	
 	<?php
-	if(function_exists('bcn_display') && !is_front_page()) {
+	if(function_exists('bcn_display') && !is_front_page() && !is_page(777)) {
 		echo '<div class="breadcrumbs" typeof="BreadcrumbList" vocab="https://schema.org/">';
 		bcn_display();
 		echo '</div>';
 	}?>
-	<div id="content" class="site-content">
+	
+	<?php if (is_page(777)) { ?>
+		<div id="custom-content" class="custom-site-content">
+	<?php } else { ?>
+		<div id="content" class="site-content">
+	<?php } ?>
